Refatorar - Long para double

+++ src/main/java/br/edu/ifg/figuras/FiguraGeometrica.java
@@ -5,6 +5,6 @@ package br.edu.ifg.figuras;
  * @author Victor
  */
 public abstract class FiguraGeometrica {
-    public abstract Long getArea();
-    public abstract Long getPerimetro();
+    public abstract double getArea();
+    public abstract double getPerimetro();
 }
